requirements.txt
anyio==3.3.1
asgiref==3.3.4
bcrypt==3.2.0
beautifulsoup4==4.9.3
blackpink==0.6
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
cloudscraper==1.2.58
colorama==0.4.4
commonmark==0.9.1
cryptography==3.4.8
decorator==5.0.9
Django==3.2.4
django-cors-headers==3.8.0
djongo==1.3.6
emoji==1.4.2
filetype==1.0.7
flask
h11==0.12.0
html2text==2020.1.16
httpcore==0.13.7
httpx==0.19.0
humanfriendly==9.2
idna==2.10
Js2Py==0.71
lxml==4.6.3
multicolorcaptcha==1.1.3
oauthlib==3.1.1
otakudesu==0.0.9
pafy==0.5.5
phlogo==0.6
Pillow==8.4.0
psutil==5.8.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
pycryptodomex==3.9.9
pydantic==1.8.2
Pygments==2.10.0
pyjsparser==2.7.1
PyJWT==2.3.0
pymongo==3.12.0
pyparsing==2.4.7
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==10.12.0
saucenao-api==2.3.1
simplejson==3.17.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.2.4
tweepy==3.10.0
twilio==7.2.0
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.7
validators==0.18.2
youtube-dl==2021.4.1
youtube-search-python==1.4.6
